Iex.jl Documentation

A Julia wrapper for the IEX Cloud API.

Overview

This package is a Julia wrapper for the IEX Cloud API. IEX Cloud provides realtime and historical data for commodities, equities, digital currencies (i.e. cryptocurrencies), foreign exchange etc.

The pricing structure is flexible and scalable, depending on usage measured in terms of messages/month. Each API call has a pre-specified message weight which is debited from the current available message quota for the month.

The free tier level provides 50,000 core messages/month, unlimited Investors Exchange data, unlimited sandbox testing, and limited access to core data. A new account can be established here.

Installation

Iex.balance_sheet — Method

balance_sheet(conn::Connection, symbol::String; period::String="quarter",
    last::Int=1)

Pulls balance sheet data. Available quarterly or annually with the default being the last available quarter.

Iex.batch — Method

batch(conn::Connection, symbol::String, types::String; symbols="",
    daterange="1m",
    columns = "date,open,high,low,close,volume,uOpen,uHigh,uLow,uClose,uVolume"

)

Batch various API endpoints for a symbol or list of symbols

Path Parameters

Option	Description

symbol	Use 'market' to get_stocks multiple symbols (i.e. .../market/batch?...)

Query String Parameters

Option	Details

types	Required

Comma delimited list of endpoints to call. The names should match

the individual endpoint names. Limited to 10 endpoints.

symbols	Optional

Comma delimited list of symbols limited to 100. This parameter is

used only if market option is used.

***(NOTE: If you do not include a symbols parameter the default

behavior is to return the data for ALL relevant symbols. Please be

cautious when using this behavior as it can drive up message usage

quickly.)***

range	Optional

Used to specify a chart range if chart is used in types parameter.

*	Optional

Parameters that are sent to individual endpoints can be specified

in batch calls and will be applied to each supporting endpoint.

For example, last can be used for the news endpoint to specify the

number of articles

Iex.book — Method

book(conn::Connection, symbol::String)

Response includes data from deep and quote. Refer to each endpoint for details.

Iex.cashflow — Method

cashflow(conn::Connection, symbol::String; period::String="quarter",
    last::Int=get_iex1)

Pulls cash flow data. Available quarterly or annually, with the default being the last available quarter.

Iex.company — Method

profile(conn::Connection, symbol::String)

Get company profile

Iex.dividends — Method

dividends(conn::Connection, symbol::String; daterange::String="1m")

Provides basic dividend data for US equities, ETFs, and Mutual Funds for the last 5 years. For 13+ years of history and comprehensive data, use the Advanced Dividends endpoint.

Iex.earnings — Method

earnings(conn::Connection, symbol::String; period::String="quarter",
    field::String="actualEPS", last::Int=1)

Earnings data for a given company including the actual EPS, consensus, and fiscal period. Earnings are available quarterly (last 4 quarters) and annually (last 4 years).

Iex.get_stocks — Method

get_stocks(conn::Connection, endpoint::String, path_params;
    query_params=Dict())

Get stock prices in various formats

Iex.history — Method

history(conn::Connection, symbol::String; daterange="1m",
    chartcloseonly::Bool = false,
    columns = "date,open,high,low,close,volume,uOpen,uHigh,uLow,uClose,uVolume"
)

Gets adjusted and unadjusted historical data for up to 15 years. Useful for building charts.

This endpoint is carried over from the IEX 1.0 API, and we know it has become difficult to use. We are working on new endpoints to simplify time series queries.

Iex.income — Method

income(conn::Connection, symbol::String; period::String="quarter",
    last::Int=1)

Earnings data for a given company including the actual EPS, consensus, and fiscal period. Earnings are available quarterly (last 4 quarters) and annually (last 4 years).

Iex.init_connection — Method

init_connection(;url::String = "",version::String="",token::String="")

initializes IEX connection parameters

Iex.insider_roster — Method

insider_roster(conn::Connection, symbol::String)

Returns the top 10 insiders, with the most recent information.

Iex.insider_summary — Method

insider_summary(conn::Connection, symbol::String)

Returns aggregated insiders summary data for the last 6 months.

Iex.insider_transactions — Method

insider_transactions(conn::Connection, symbol::String)

Returns insider transactions.

Iex.logo — Method

logo(conn::Connection, symbol::String)

Get company logo

Iex.metadata — Method

metadata(conn::Connection)

Get metadata for account associated wth IEX connection conn

Iex.news — Method

news(conn::Connection, symbol::String; last::Int=10

Provides intraday news from over 3,000 global news sources including major publications, regional media, and social.. This endpoint returns up to the last 50 articles. Use the historical news endpoint to fetch news as far back as January 2019

Path Parameters: |Option|Description |:––-|:––––– |symbol|A stock symbol |last |Number between 1 and 50. Default is 10

Iex.ohlc — Method

ohlc(conn::Connection, symbol::String)

Response includes data from deep and quote. Refer to each endpoint for details.

Iex.peers — Method

peers(conn::Connection, symbol::String)

Get company peer group

Iex.price — Method

price(conn::Connection, symbol::String)

Response includes data from deep and quote. Refer to each endpoint for details.

Iex.set_connection_token! — Method

set_connection_token!(conn::Connection, token::String)

Modifies url for connection object conn

Iex.set_connection_url! — Method

set_connection_url!(conn::Connection, url::String)

Modifies url for connection object conn

Iex.set_connection_version! — Method

set_connection_version!(conn::Connection, version::String)

Modifies version for connection object conn

Iex.symbol_quote — Method

symbol_quote(conn::Connection, symbol::String; field::String="latestPrice")

Iex.symbols — Method

symbols(conn::Connection; format::String="JSON")

This call returns an array of symbols the Investors Exchange supports for trading. This list is updated daily as of 7:45 a.m. ET. Symbols may be added or removed by the Investors Exchange after the list was produced

Iex.usage — Method

usage(get_iex; quota_type="messages")

Get usage for account associated with IEX connection conn

Iex.Connect.IEXException — Type

struct IEXException <: Exception

IEX Cloud uses HTTP response codes to indicate the success or failure of an API request.

Found at API Version

Iex.Connect.status — Method

status(conn::Connection)

Get status for IEX connection

# Iex.jl is not currently registered as an official package
# Please install the development version from GitHub:

Pkg.clone("git://GitHub.com/GodolphinCapitalManagement/Iex.jl")

If you encounter a clear bug, please file a minimal reproducible example on GitHub.

Functions

Time Series for AAPL

Iex.history(iex, "AAPL")
Iex.symbol_quote(iex, "AAPL")

Usage

using DataFrames
using Plots

using Iex
using Iex.Connect

params = (
    token=ENV["IEX_SANDBOX_API_KEY"],
    url="https://sandbox.iexapis.com",
    version="stable",
)
iex = init_connection(;url=params.url, version=params.version, token=params.token);

gr(fmt=:png, size=(800,470))
# Get daily S&P 500 data
data = history(iex, "SPY");

# Plot the timeseries
plot(data.date, data.close, label="Close", title="SPY")
plot!(data.date, data.high, label="High")
plot!(data.date, data.low, label="Low")
savefig("spy.png")